chore: Schematics - dependency migration (#12522)
closes #12044
