+++ projects/schematics/src/dependencies.json
@@ -51,6 +51,153 @@
     "@spartacus/core": "3.3.0",
     "@spartacus/storefront": "3.3.0"
   },
+  "@spartacus/asm": {
+    "@angular-devkit/schematics": "^10.2.1",
+    "@angular/common": "^10.2.4",
+    "@angular/core": "^10.2.4",
+    "@angular/forms": "^10.2.4",
+    "@ngrx/effects": "^10.0.0",
+    "@ngrx/store": "^10.0.0",
+    "@spartacus/core": "3.2.0-next.2",
+    "@spartacus/schematics": "3.2.0-next.2",
+    "@spartacus/storefront": "3.2.0-next.2",
+    "rxjs": "^6.6.0"
+  },
+  "@spartacus/cart": {
+    "@angular-devkit/schematics": "^10.1.0",
+    "@angular/common": "^10.1.0",
+    "@angular/core": "^10.1.0",
+    "@angular/forms": "^10.1.0",
+    "@angular/router": "^10.1.0",
+    "@ngrx/effects": "^10.0.0",
+    "@ngrx/store": "^10.0.0",
+    "@spartacus/core": "3.2.0-next.2",
+    "@spartacus/schematics": "3.2.0-next.2",
+    "@spartacus/storefront": "3.2.0-next.2",
+    "@spartacus/styles": "3.2.0-next.2",
+    "bootstrap": "^4.0",
+    "rxjs": "^6.6.0"
+  },
+  "@spartacus/organization": {
+    "@angular-devkit/schematics": "^10.1.0",
+    "@angular/common": "^10.1.0",
+    "@angular/core": "^10.1.0",
+    "@angular/forms": "^10.1.0",
+    "@angular/router": "^10.1.0",
+    "@ng-select/ng-select": "^5.0.9",
+    "@ngrx/effects": "^10.0.0",
+    "@ngrx/store": "^10.0.0",
+    "@spartacus/core": "3.2.0-next.2",
+    "@spartacus/schematics": "3.2.0-next.2",
+    "@spartacus/storefront": "3.2.0-next.2",
+    "bootstrap": "^4.0",
+    "rxjs": "^6.6.0"
+  },
+  "@spartacus/product": {
+    "@angular-devkit/schematics": "^10.1.0",
+    "@angular/common": "^10.1.0",
+    "@angular/core": "^10.1.0",
+    "@angular/router": "^10.1.0",
+    "@spartacus/core": "3.2.0-next.2",
+    "@spartacus/schematics": "3.2.0-next.2",
+    "@spartacus/storefront": "3.2.0-next.2",
+    "bootstrap": "^4.0",
+    "rxjs": "^6.6.0"
+  },
+  "@spartacus/product-configurator": {
+    "@angular-devkit/schematics": "^10.2.1",
+    "@angular/common": "^10.2.4",
+    "@angular/core": "^10.2.4",
+    "@angular/forms": "^10.2.4",
+    "@angular/router": "^10.2.4",
+    "@ng-select/ng-select": "^5.0.9",
+    "@ngrx/effects": "^10.0.0",
+    "@ngrx/store": "^10.0.0",
+    "@spartacus/core": "3.2.0-next.2",
+    "@spartacus/schematics": "3.2.0-next.2",
+    "@spartacus/storefront": "3.2.0-next.2",
+    "@spartacus/styles": "3.2.0-next.2",
+    "rxjs": "^6.6.0"
+  },
+  "@spartacus/qualtrics": {
+    "@angular-devkit/schematics": "^10.1.0",
+    "@angular/common": "^10.1.0",
+    "@angular/core": "^10.1.0",
+    "@spartacus/core": "3.2.0-next.2",
+    "@spartacus/schematics": "3.2.0-next.2",
+    "bootstrap": "^4.0",
+    "rxjs": "^6.6.0"
+  },
+  "@spartacus/smartedit": {
+    "@angular-devkit/schematics": "^10.1.0",
+    "@angular/common": "^10.1.0",
+    "@angular/core": "^10.1.0",
+    "@spartacus/core": "3.2.0-next.2",
+    "@spartacus/schematics": "3.2.0-next.2",
+    "rxjs": "^6.6.0"
+  },
+  "@spartacus/storefinder": {
+    "@angular-devkit/schematics": "^10.1.0",
+    "@angular/common": "^10.1.0",
+    "@angular/core": "^10.1.0",
+    "@angular/forms": "^10.1.0",
+    "@angular/router": "^10.1.0",
+    "@ng-bootstrap/ng-bootstrap": "^7.0.0",
+    "@ngrx/effects": "^10.0.0",
+    "@ngrx/store": "^10.0.0",
+    "@spartacus/core": "3.2.0-next.2",
+    "@spartacus/schematics": "3.2.0-next.2",
+    "@spartacus/storefront": "3.2.0-next.2",
+    "bootstrap": "^4.3.1",
+    "rxjs": "^6.6.0"
+  },
+  "@spartacus/tracking": {
+    "@angular-devkit/schematics": "^10.2.1",
+    "@angular/common": "^10.2.4",
+    "@angular/core": "^10.1.0",
+    "@spartacus/core": "3.2.0-next.2",
+    "@spartacus/schematics": "3.2.0-next.2",
+    "rxjs": "^6.6.0"
+  },
+  "@spartacus/user": {
+    "@angular-devkit/schematics": "^10.2.1",
+    "@angular/common": "^10.1.0",
+    "@angular/core": "^10.1.0",
+    "@angular/forms": "^10.1.0",
+    "@angular/router": "^10.1.0",
+    "@ngrx/store": "^10.0.0",
+    "@spartacus/core": "3.2.0-next.2",
+    "@spartacus/schematics": "3.2.0-next.2",
+    "@spartacus/storefront": "3.2.0-next.2",
+    "@spartacus/styles": "3.2.0-next.2",
+    "rxjs": "^6.6.0"
+  },
+  "@spartacus/cdc": {
+    "@angular-devkit/schematics": "^10.2.1",
+    "@angular/common": "^10.1.0",
+    "@angular/core": "^10.1.0",
+    "@angular/router": "^10.1.0",
+    "@ngrx/effects": "^10.0.0",
+    "@ngrx/store": "^10.0.0",
+    "@spartacus/asm": "3.2.0-next.2",
+    "@spartacus/core": "3.2.0-next.2",
+    "@spartacus/schematics": "3.2.0-next.2",
+    "@spartacus/storefront": "3.2.0-next.2",
+    "@spartacus/user": "3.2.0-next.2",
+    "rxjs": "^6.6.0"
+  },
+  "@spartacus/cds": {
+    "@angular-devkit/schematics": "^10.2.1",
+    "@angular/common": "^10.1.0",
+    "@angular/core": "^10.1.0",
+    "@angular/router": "^10.1.0",
+    "@ngrx/store": "^10.0.0",
+    "@spartacus/core": "3.2.0-next.2",
+    "@spartacus/schematics": "3.2.0-next.2",
+    "@spartacus/storefront": "3.2.0-next.2",
+    "@spartacus/tracking": "3.2.0-next.2",
+    "rxjs": "^6.6.0"
+  },
   "storefrontapp": {
     "@angular/animations": "^10.2.4",
     "@angular/common": "^10.2.4",
